DD-796 move env var function to separate file
